Only remove images if there was a new one committed
Before, it looks like it was removing images
but sometimes not committing a new image.
Especially at the end of the build,
when it's really important that there be
1 image available.
Sometimes there was no image,
so it couldn't open the finalTerminal.
Also, don't clean up the images
on the timeout.
There's already a function that does that
when closing all terminals.

+++ src/utils/commitContainer.ts
@@ -20,14 +20,15 @@ export default function commitContainer(
 
       running_container=$(get_running_container ${dockerImage})
       if [[ -n $running_container ]]; then
-        latest_committed_image=$(docker images ${imageRepo} --format "{{.ID}} {{.Tag}}" | sort -k 2 -h | tail -n1 | awk '{print $1}')
-        for previous_image in $(docker images -q "${imageRepo}"); do
-          if [[ $previous_image != $latest_committed_image ]]; then
-            docker rmi $previous_image
-          fi
-        done;
-
-        docker commit --pause=false $running_container ${newImageRepoAndTag}
+        committed_image=$(docker commit --pause=false $running_container ${newImageRepoAndTag})
+        if [[ -n $committed_image ]]; then
+          latest_committed_image=$(docker images ${imageRepo} --format "{{.ID}} {{.Tag}}" | sort -k 2 -h | tail -n1 | awk '{print $1}')
+          for previous_image in $(docker images -q "${imageRepo}"); do
+            if [[ $previous_image != $latest_committed_image ]]; then
+              docker rmi $previous_image
+            fi
+          done;
+        fi
       fi`,
     ],
     getSpawnOptions()
